Resolved the db query issues Now need to convert those data into listing
Now know how to get any data with query in db. But now have to make these data readable and usable with mysqli functions., have to figure out which functions or class to be used so that we get the list with Id wise hobbies. The problem is - we have to separate the arrays by unique id to get the list of hobbies by user_id